Update deploy script to deploy only if build succeeded.

+++ deploy.sh
@@ -9,12 +9,20 @@ DEPLOY_BRANCH="master"
 # Only deploy from master branch.
 CURRENT_BRANCH="$(git rev-parse --abbrev-ref HEAD)"
 if [[ ${CURRENT_BRANCH} == ${DEPLOY_BRANCH} ]]; then
+	# Stash any uncommitted changes.
 	STASH_SUCCEEDED=false
 	if ! git stash --include-untracked | grep "No local changes to save"; then
 	  STASH_SUCCEEDED=true
 	fi
-	npm run build
-	rsync -avz --delete ${DEPLOY_SRC_DIR}/ ${DEPLOY_USERNAME}@${DEPLOY_HOST}:${DEPLOY_DEST_DIR}/
+
+	# Only deploy if build succeeded.
+	if npm run build; then
+		rsync -avz --delete ${DEPLOY_SRC_DIR}/ ${DEPLOY_USERNAME}@${DEPLOY_HOST}:${DEPLOY_DEST_DIR}/
+	else
+		echo "== Error: build failed; not deploying"
+	fi
+
+	# Unstash if we stashed uncommitted changes above.
 	if [[ ${STASH_SUCCEEDED} == true ]]; then
 	  git stash pop
 	fi
